ray and sphere example

+++ src/maths/maths.h
@@ -6,7 +6,7 @@
 /*   By: liguyon <liguyon@student.42lehavre.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/19 15:00:27 by liguyon           #+#    #+#             */
-/*   Updated: 2024/01/20 01:32:22 by liguyon          ###   ########.fr       */
+/*   Updated: 2024/01/20 04:02:59 by liguyon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ t_vec3	vec3_div(t_vec3 v, float scalar);
 float	vec3_length_squared(t_vec3 v);
 float	vec3_length(t_vec3 v);
 t_vec3	vec3_normalize(t_vec3 v);
-double	vec3_dot(t_vec3 v, t_vec3 w);
+float	vec3_dot(t_vec3 v, t_vec3 w);
 t_vec3	vec3_cross(t_vec3 v, t_vec3 w);
 
 /* Misc
