+++ src/maths/vec3_1.c
@@ -6,7 +6,7 @@
 /*   By: liguyon <liguyon@student.42lehavre.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/23 14:39:00 by liguyon           #+#    #+#             */
-/*   Updated: 2024/01/19 22:34:01 by liguyon          ###   ########.fr       */
+/*   Updated: 2024/01/20 04:02:45 by liguyon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ t_vec3	vec3_normalize(t_vec3 v)
 	return (ret);
 }
 
-double	vec3_dot(t_vec3 v, t_vec3 w)
+float	vec3_dot(t_vec3 v, t_vec3 w)
 {
 	return (v.x * w.x + v.y * w.y + v.z * w.z);
 }
